LPS-201521 Extract test logic of importing and exporting ObjectDefinitions to a base class
