--- modules/apps/object/object-test/src/testIntegration/java/com/liferay/object/web/internal/portlet/action/test/ExportImportObjectDefinitionTest.java
@@ -1,287 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.object.web.internal.portlet.action.test;
-
-import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
-import com.liferay.object.admin.rest.dto.v1_0.ObjectDefinition;
-import com.liferay.object.admin.rest.resource.v1_0.ObjectDefinitionResource;
-import com.liferay.petra.string.StringBundler;
-import com.liferay.petra.string.StringPool;
-import com.liferay.portal.kernel.model.Layout;
-import com.liferay.portal.kernel.model.LayoutConstants;
-import com.liferay.portal.kernel.model.User;
-import com.liferay.portal.kernel.portlet.LiferayPortletRequest;
-import com.liferay.portal.kernel.portlet.bridges.mvc.MVCActionCommand;
-import com.liferay.portal.kernel.portlet.bridges.mvc.MVCResourceCommand;
-import com.liferay.portal.kernel.test.ReflectionTestUtil;
-import com.liferay.portal.kernel.test.portlet.MockLiferayPortletActionRequest;
-import com.liferay.portal.kernel.test.portlet.MockLiferayPortletActionResponse;
-import com.liferay.portal.kernel.test.portlet.MockLiferayResourceRequest;
-import com.liferay.portal.kernel.test.portlet.MockLiferayResourceResponse;
-import com.liferay.portal.kernel.test.util.RandomTestUtil;
-import com.liferay.portal.kernel.test.util.TestPropsValues;
-import com.liferay.portal.kernel.theme.ThemeDisplay;
-import com.liferay.portal.kernel.util.ArrayUtil;
-import com.liferay.portal.kernel.util.File;
-import com.liferay.portal.kernel.util.LocaleUtil;
-import com.liferay.portal.kernel.util.Portal;
-import com.liferay.portal.kernel.util.ProxyUtil;
-import com.liferay.portal.kernel.util.StringUtil;
-import com.liferay.portal.kernel.util.Validator;
-import com.liferay.portal.kernel.util.WebKeys;
-import com.liferay.portal.model.impl.LayoutImpl;
-import com.liferay.portal.test.rule.Inject;
-import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
-import com.liferay.portal.upload.test.util.UploadTestUtil;
-import com.liferay.portal.vulcan.pagination.Page;
-import com.liferay.portal.vulcan.pagination.Pagination;
-
-import java.util.List;
-import java.util.Objects;
-
-import org.junit.Before;
-import org.junit.ClassRule;
-import org.junit.Rule;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-
-import org.skyscreamer.jsonassert.JSONAssert;
-import org.skyscreamer.jsonassert.JSONCompareMode;
-
-import org.springframework.http.MediaType;
-import org.springframework.mock.web.MockMultipartFile;
-import org.springframework.mock.web.MockMultipartHttpServletRequest;
-
-/**
- * @author Gabriel Albuquerque
- */
-@RunWith(Arquillian.class)
-public class ExportImportObjectDefinitionTest {
-
-	@ClassRule
-	@Rule
-	public static final LiferayIntegrationTestRule liferayIntegrationTestRule =
-		new LiferayIntegrationTestRule();
-
-	@Before
-	public void setUp() throws Exception {
-		_user = TestPropsValues.getUser();
-
-		ObjectDefinitionResource.Builder builder =
-			_objectDefinitionResourceFactory.create();
-
-		_objectDefinitionResource = builder.user(
-			_user
-		).build();
-	}
-
-	@Test
-	public void testExportImportObjectDefinition() throws Exception {
-		_testExportImportObjectDefinition(
-			"account_entry_system_object_definition.json", "AccountEntry",
-			true);
-		_testExportImportObjectDefinition(
-			"custom_object_definition.json", "ImportedCustomObjectDefinition",
-			false);
-	}
-
-	private MockLiferayPortletActionRequest
-			_createMockLiferayPortletActionRequest(
-				String externalReferenceCode, String fileName, String name)
-		throws Exception {
-
-		MockLiferayPortletActionRequest mockLiferayPortletActionRequest =
-			new MockLiferayPortletActionRequest(
-				_createMockMultipartHttpServletRequest(fileName));
-
-		if (Validator.isNotNull(externalReferenceCode)) {
-			mockLiferayPortletActionRequest.addParameter(
-				"externalReferenceCode", externalReferenceCode);
-		}
-
-		mockLiferayPortletActionRequest.addParameter("name", name);
-		mockLiferayPortletActionRequest.addParameter(
-			"redirect", RandomTestUtil.randomString());
-		mockLiferayPortletActionRequest.setAttribute(
-			WebKeys.THEME_DISPLAY, _getThemeDisplay());
-
-		return mockLiferayPortletActionRequest;
-	}
-
-	private MockLiferayResourceRequest _createMockLiferayResourceRequest(
-		long objectDefinitionId) {
-
-		MockLiferayResourceRequest mockLiferayResourceRequest =
-			new MockLiferayResourceRequest();
-
-		mockLiferayResourceRequest.addParameter(
-			"objectDefinitionId", String.valueOf(objectDefinitionId));
-
-		ThemeDisplay themeDisplay = new ThemeDisplay();
-
-		themeDisplay.setUser(_user);
-
-		mockLiferayResourceRequest.setAttribute(
-			WebKeys.THEME_DISPLAY, themeDisplay);
-
-		return mockLiferayResourceRequest;
-	}
-
-	private MockMultipartHttpServletRequest
-			_createMockMultipartHttpServletRequest(String fileName)
-		throws Exception {
-
-		MockMultipartHttpServletRequest mockMultipartHttpServletRequest =
-			new MockMultipartHttpServletRequest();
-
-		Class<?> clazz = getClass();
-
-		byte[] bytes = _file.getBytes(
-			clazz.getResourceAsStream("dependencies/" + fileName));
-
-		mockMultipartHttpServletRequest.addFile(
-			new MockMultipartFile(fileName, bytes));
-
-		mockMultipartHttpServletRequest.setCharacterEncoding(StringPool.UTF8);
-
-		String boundary = "WebKitFormBoundary" + StringUtil.randomString();
-
-		mockMultipartHttpServletRequest.setContent(
-			_getContent(boundary, bytes, fileName));
-		mockMultipartHttpServletRequest.setContentType(
-			MediaType.MULTIPART_FORM_DATA_VALUE + "; boundary=" + boundary);
-
-		return mockMultipartHttpServletRequest;
-	}
-
-	private byte[] _getContent(String boundary, byte[] bytes, String fileName) {
-		String start = StringBundler.concat(
-			StringPool.DOUBLE_DASH, boundary,
-			"\r\nContent-Disposition:form-data;",
-			"name=\"objectDefinitionJSON\";filename=\"", fileName,
-			"\";\r\nContent-type:application/json\r\n\r\n");
-		String end = StringBundler.concat(
-			"\r\n--", boundary, StringPool.DOUBLE_DASH);
-
-		return ArrayUtil.append(start.getBytes(), bytes, end.getBytes());
-	}
-
-	private ObjectDefinition _getObjectDefinition(String name)
-		throws Exception {
-
-		Page<ObjectDefinition> page =
-			_objectDefinitionResource.getObjectDefinitionsPage(
-				name, null, null, Pagination.of(1, 1), null);
-
-		List<ObjectDefinition> items = (List<ObjectDefinition>)page.getItems();
-
-		return items.get(0);
-	}
-
-	private ThemeDisplay _getThemeDisplay() throws Exception {
-		ThemeDisplay themeDisplay = new ThemeDisplay();
-
-		Layout layout = new LayoutImpl();
-
-		layout.setType(LayoutConstants.TYPE_CONTROL_PANEL);
-
-		themeDisplay.setLayout(layout);
-
-		themeDisplay.setScopeGroupId(TestPropsValues.getGroupId());
-		themeDisplay.setSiteDefaultLocale(LocaleUtil.US);
-		themeDisplay.setUser(_user);
-
-		return themeDisplay;
-	}
-
-	private void _testExportImportObjectDefinition(
-			String fileName, String objectDefinitionName, boolean system)
-		throws Exception {
-
-		String externalReferenceCode = null;
-		ObjectDefinition objectDefinition = null;
-
-		if (system) {
-			objectDefinition = _getObjectDefinition(objectDefinitionName);
-
-			externalReferenceCode = objectDefinition.getExternalReferenceCode();
-		}
-
-		MockLiferayPortletActionRequest mockLiferayPortletActionRequest =
-			_createMockLiferayPortletActionRequest(
-				externalReferenceCode, fileName, objectDefinitionName);
-
-		ReflectionTestUtil.setFieldValue(
-			_mvcActionCommand, "_portal",
-			ProxyUtil.newProxyInstance(
-				ExportImportObjectDefinitionTest.class.getClassLoader(),
-				new Class<?>[] {Portal.class},
-				(proxy, method, args) -> {
-					if (Objects.equals(
-							method.getName(), "getUploadPortletRequest")) {
-
-						LiferayPortletRequest liferayPortletRequest =
-							_portal.getLiferayPortletRequest(
-								mockLiferayPortletActionRequest);
-
-						return UploadTestUtil.createUploadPortletRequest(
-							_portal.getUploadServletRequest(
-								liferayPortletRequest.getHttpServletRequest()),
-							liferayPortletRequest,
-							_portal.getPortletNamespace(
-								liferayPortletRequest.getPortletName()));
-					}
-
-					return method.invoke(_portal, args);
-				}));
-
-		_mvcActionCommand.processAction(
-			mockLiferayPortletActionRequest,
-			new MockLiferayPortletActionResponse());
-
-		objectDefinition = _getObjectDefinition(objectDefinitionName);
-
-		MockLiferayResourceResponse mockLiferayResourceResponse =
-			new MockLiferayResourceResponse();
-
-		_mvcResourceCommand.serveResource(
-			_createMockLiferayResourceRequest(objectDefinition.getId()),
-			mockLiferayResourceResponse);
-
-		Class<?> clazz = getClass();
-
-		JSONAssert.assertEquals(
-			StringUtil.read(
-				clazz.getResourceAsStream("dependencies/" + fileName)),
-			String.valueOf(
-				mockLiferayResourceResponse.getPortletOutputStream()),
-			JSONCompareMode.STRICT_ORDER);
-	}
-
-	@Inject
-	private File _file;
-
-	@Inject(
-		filter = "mvc.command.name=/object_definitions/import_object_definition"
-	)
-	private MVCActionCommand _mvcActionCommand;
-
-	@Inject(
-		filter = "mvc.command.name=/object_definitions/export_object_definition"
-	)
-	private MVCResourceCommand _mvcResourceCommand;
-
-	private ObjectDefinitionResource _objectDefinitionResource;
-
-	@Inject
-	private ObjectDefinitionResource.Factory _objectDefinitionResourceFactory;
-
-	@Inject
-	private Portal _portal;
-
-	private User _user;
-
-}
